Changed Variable from Int to Double

+++ app/src/main/java/com/roomdbwithmigration/model/Student.kt
@@ -8,5 +8,6 @@ data class Student(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val name: String,
     val age: Int,
-    val marks: Int
+    //val marks: Int
+    val marks: Double
 )
